Separated the menu system
Separated the menu system so that the mouse position is relative to the
menu itself as opposed to the whole screen as well as the drawing.  Now
the paint method creates a surface, all of the data of the menu is drawn
on the surface, and then that surface is placed on the original.